requirements.txt
# OTPP Data Processing Pipeline Requirements

# Core Python packages
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0

# Web scraping and automation
selenium>=4.15.0
undetected-chromedriver>=3.5.0
requests>=2.31.0

# PDF processing
PyMuPDF>=1.23.0
pdf2docx>=0.5.6

# Document processing  
python-docx>=0.8.11

# Date and time utilities
python-dateutil>=2.8.2

# Optional: Progress bars and CLI enhancements
tqdm>=4.66.0

# Optional: Logging enhancements
colorlog>=6.7.0

# Development and testing (optional)
pytest>=7.4.0
pytest-cov>=4.1.0

# Data validation (optional)
pydantic>=2.4.0

# Configuration management (optional)
python-decouple>=3.8